Add files via upload
The inputs are the values of current drawn by load. Power loss in transmission line is calculated using current*current*resistance. We have assumed that all the loads are at equal distance so resistance across each load will be same. Using this we can calculate the total current and we can calculate the losses. The value of resistance is very small so I used scientific notation for it. This code is 100% error free and 100% perfect. Just let me know whatever you like to know more about this code through normal message. I will reply you before tomorrow evening because I won't be able to get up so early.